+++ src/components/Menus/Sidebar/InstructionsPanel/ProgressBar/ProgressBar.test.js
@@ -6,16 +6,16 @@ import { setCurrentStepPosition } from "../../../../../redux/InstructionsSlice";
 
 let store;
 
-const renderProgressBarOnStep = (stepNumber) => {
+const renderProgressBarOnStep = (stepNumber, numberOfSteps = 3) => {
   const mockStore = configureStore([]);
+  const steps = new Array(numberOfSteps).fill(0).map((i) => ({
+    content: `<p>step ${i + 1}</p>`,
+  }));
+
   const initialState = {
     instructions: {
       project: {
-        steps: [
-          { content: "<p>step 0</p>" },
-          { content: "<p>step 1</p>" },
-          { content: "<p>step 2</p>" },
-        ],
+        steps,
       },
       currentStepPosition: stepNumber,
     },
@@ -27,6 +27,24 @@ const renderProgressBarOnStep = (stepNumber) => {
     </Provider>,
   );
 };
+
+describe("When there is only one step", () => {
+  beforeEach(() => {
+    renderProgressBarOnStep(0, 1);
+  });
+
+  test("Previous step button is not shown", () => {
+    expect(
+      screen.queryByTitle("instructionsPanel.previousStep"),
+    ).not.toBeInTheDocument();
+  });
+
+  test("Next step button is not shown", () => {
+    expect(
+      screen.queryByTitle("instructionsPanel.nextStep"),
+    ).not.toBeInTheDocument();
+  });
+});
 
 describe("When on first step", () => {
   beforeEach(() => {