+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Editable instructions
 - Support for the `outputPanels` attribute in the `PyodideRunner` (#1157)
 
 ## [0.28.14] - 2025-01-06